build: 添加构建脚本和配置优化
- 新增 .prettierrc 配置文件统一代码风格
- 添加 archive.js 和 snapshot.js 脚本用于构建产物打包
- 修改 gh-pages.sh 脚本支持子目录部署
- 更新 package.json 移除 type: module 并调整依赖顺序
- 优化 vite.config.ts 添加路径别名和配置项
