+++ package.json
@@ -3,7 +3,6 @@
   "version": "1.0.1",
   "author": "gausszhou",
   "private": true,
-  "type": "module",
   "scripts": {
     "dev": "vite --host",
     "build": "rm -rf dist && tsc && vite build",
@@ -11,13 +10,23 @@
   },
   "dependencies": {
     "@ant-design/icons": "^6.0.0",
-    "@types/uuid": "^10.0.0",
     "antd": "^5.25.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "svgo": "^3.3.2",
     "upng-js": "^2.1.0",
     "uuid": "^11.1.0"
+  },
+  "devDependencies": {
+    "@types/node": "^24.9.2",
+    "@types/react": "^19.1.3",
+    "@types/react-dom": "^19.1.3",
+    "@types/uuid": "^10.0.0",
+    "@vitejs/plugin-react": "^4.4.1",
+    "adm-zip": "^0.5.12",
+    "dotenv": "^16.5.0",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.5"
   },
   "eslintConfig": {
     "extends": [
@@ -36,14 +45,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/react": "^19.1.3",
-    "@types/react-dom": "^19.1.3",
-    "@vitejs/plugin-react": "^4.4.1",
-    "adm-zip": "^0.5.12",
-    "dotenv": "^16.5.0",
-    "typescript": "^5.8.3",
-    "vite": "^6.3.5"
   }
 }
